# api/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"post":{"operationId":"AppController_create","summary":"","description":"Returns a stringified version of the CreateUserDto object.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/register":{"post":{"operationId":"AuthController_register","summary":"","description":"Register a new user and return an access token.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDto"}}}},"422":{"description":"Email already exists."}}}},"/auth/signin":{"post":{"operationId":"AuthController_signIn","summary":"","description":"Sign in with an email and password and return an access token.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDto"}}}},"401":{"description":"Invalid email or password."}}}}},"info":{"title":"Fullstack Starter API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","description":"This comment appears in the generated Swagger documentation."}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"fullName":{"type":"string"}},"required":["email","password","fullName"]},"AccessTokenDto":{"type":"object","properties":{"access_token":{"type":"string","description":"A JWT access token."}},"required":["access_token"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}